+++ lib/pocketbase_utils.dart
@@ -1,3 +1,1 @@
-library;
-
 export 'src/generator/generator.dart';