+++ lib/src/config/pubspec_config.dart
@@ -1,9 +1,39 @@
+import 'package:pocketbase_utils/src/config/config_exception.dart';
+import 'package:pocketbase_utils/src/utils/file_utils.dart';
 import 'package:yaml/yaml.dart' as yaml;
 
-import '../utils/file_utils.dart';
-import 'config_exception.dart';
-
 class PubspecConfig {
+  PubspecConfig() {
+    final pubspecFile = getPubspecFile();
+    if (pubspecFile == null) {
+      throw ConfigException("Can't find 'pubspec.yaml' file.");
+    }
+
+    final pubspecFileContent = pubspecFile.readAsStringSync();
+    final pubspecYaml = yaml.loadYaml(pubspecFileContent);
+    if (pubspecYaml is! yaml.YamlMap) {
+      throw ConfigException(
+        "Failed to extract config from the 'pubspec.yaml' file.\nExpected YAML map but got ${pubspecYaml.runtimeType}.",
+      );
+    }
+
+    final pocketbaseUtilsConfig = pubspecYaml['pocketbase_utils'];
+    if (pocketbaseUtilsConfig is! yaml.YamlMap) {
+      throw ConfigException(
+        "Failed to extract pocketbaseUtilsConfig from the 'pubspec.yaml' file.\nExpected YAML map but got ${pocketbaseUtilsConfig.runtimeType}.",
+      );
+    }
+
+    _enabled = pocketbaseUtilsConfig['enabled'] is bool ? pocketbaseUtilsConfig['enabled'] as bool : null;
+    _pbSchemaPath =
+        pocketbaseUtilsConfig['pb_schema_path'] is String ? pocketbaseUtilsConfig['pb_schema_path'] as String : null;
+    _outputDir = pocketbaseUtilsConfig['output_dir'] is String ? pocketbaseUtilsConfig['output_dir'] as String : null;
+    _lineLength = pocketbaseUtilsConfig['line_length'] is int ? pocketbaseUtilsConfig['line_length'] as int : null;
+    _generateSystemCollections = pocketbaseUtilsConfig['generate_system_collections'] is bool
+        ? pocketbaseUtilsConfig['generate_system_collections'] as bool
+        : null;
+  }
+
   bool? _enabled;
   String? _pbSchemaPath;
   String? _outputDir;
@@ -19,33 +49,4 @@ class PubspecConfig {
   int? get lineLength => _lineLength;
 
   bool? get generateSystemCollections => _generateSystemCollections;
-
-  PubspecConfig() {
-    var pubspecFile = getPubspecFile();
-    if (pubspecFile == null) {
-      throw ConfigException("Can't find 'pubspec.yaml' file.");
-    }
-
-    var pubspecFileContent = pubspecFile.readAsStringSync();
-    var pubspecYaml = yaml.loadYaml(pubspecFileContent);
-
-    if (pubspecYaml is! yaml.YamlMap) {
-      throw ConfigException(
-        "Failed to extract config from the 'pubspec.yaml' file.\nExpected YAML map but got ${pubspecYaml.runtimeType}.",
-      );
-    }
-
-    var pocketbaseUtilsConfig = pubspecYaml['pocketbase_utils'];
-    if (pocketbaseUtilsConfig == null) {
-      return;
-    }
-
-    _enabled = pocketbaseUtilsConfig['enabled'] is bool ? pocketbaseUtilsConfig['enabled'] : null;
-    _pbSchemaPath = pocketbaseUtilsConfig['pb_schema_path'] is String ? pocketbaseUtilsConfig['pb_schema_path'] : null;
-    _outputDir = pocketbaseUtilsConfig['output_dir'] is String ? pocketbaseUtilsConfig['output_dir'] : null;
-    _lineLength = pocketbaseUtilsConfig['line_length'] is int ? pocketbaseUtilsConfig['line_length'] : null;
-    _generateSystemCollections = pocketbaseUtilsConfig['generate_system_collections'] is bool
-        ? pocketbaseUtilsConfig['generate_system_collections']
-        : null;
-  }
 }