+++ lib/src/generator/generator.dart
@@ -2,17 +2,16 @@ import 'dart:convert';
 import 'dart:io';
 
 import 'package:path/path.dart' as path;
+import 'package:pocketbase_utils/src/config/pubspec_config.dart';
 import 'package:pocketbase_utils/src/schema/collection/collection.dart';
 import 'package:pocketbase_utils/src/templates/auth_record.dart';
 import 'package:pocketbase_utils/src/templates/base_record.dart';
 import 'package:pocketbase_utils/src/templates/date_time_json_methods.dart';
 import 'package:pocketbase_utils/src/templates/empty_values.dart';
 import 'package:pocketbase_utils/src/templates/geo_point_class.dart';
+import 'package:pocketbase_utils/src/utils/file_utils.dart';
+import 'package:pocketbase_utils/src/utils/utils.dart';
 import 'package:recase/recase.dart';
-
-import '../config/pubspec_config.dart';
-import '../utils/file_utils.dart';
-import '../utils/utils.dart';
 
 const _defaultPbSchemaPath = 'pb_schema.json';
 const _defaultOutputDir = 'lib/generated/pocketbase';
@@ -21,14 +20,9 @@ const _defaultGenerateSystemCollections = false;
 
 /// The generator of models files.
 class Generator {
-  late String _pbSchemaPath;
-  late String _outputDir;
-  late int _lineLength;
-  late bool _generateSystemCollections;
-
   /// Creates a new generator with configuration from the 'pubspec.yaml' file.
   Generator() {
-    var pubspecConfig = PubspecConfig();
+    final pubspecConfig = PubspecConfig();
 
     if (pubspecConfig.enabled != true) {
       warning(
@@ -41,13 +35,17 @@ class Generator {
 
     _outputDir = pubspecConfig.outputDir ?? _defaultOutputDir;
     if (!isValidPath(_outputDir)) {
-      warning("Config parameter 'output_dir' requires valid path value (e.g. 'lib', 'lib\\generated').");
+      warning(r"Config parameter 'output_dir' requires valid path value (e.g. 'lib', 'lib\generated').");
     }
 
     _lineLength = pubspecConfig.lineLength ?? _defaultLineLength;
 
     _generateSystemCollections = pubspecConfig.generateSystemCollections ?? _defaultGenerateSystemCollections;
   }
+  late String _pbSchemaPath;
+  late String _outputDir;
+  late int _lineLength;
+  late bool _generateSystemCollections;
 
   /// Generates collections models files.
   Future<void> generateAsync() async {
@@ -57,9 +55,9 @@ class Generator {
   }
 
   Future<File> _checkPbSchemaPath() async {
-    var rootDirPath = getRootDirectoryPath();
-    var absPbSchemaPath = path.join(rootDirPath, _pbSchemaPath);
-    var pbSchemaFile = File(absPbSchemaPath);
+    final rootDirPath = getRootDirectoryPath();
+    final absPbSchemaPath = path.join(rootDirPath, _pbSchemaPath);
+    final pbSchemaFile = File(absPbSchemaPath);
 
     if (!pbSchemaFile.existsSync()) {
       exitWithError("File in absolute path $absPbSchemaPath doesn't exist!");
@@ -69,18 +67,25 @@ class Generator {
   }
 
   Future<Directory> _updateGeneratedDir() async {
-    var rootDirPath = getRootDirectoryPath();
-    var outputDirPath = path.join(rootDirPath, _outputDir);
+    final rootDirPath = getRootDirectoryPath();
+    final outputDirPath = path.join(rootDirPath, _outputDir);
 
     return Directory(outputDirPath).create(recursive: true);
   }
 
   Future<void> _generateDartFiles(File pbSchemaFile, Directory outputDirectory) async {
     final pbSchemaFileContent = pbSchemaFile.readAsStringSync();
-    final List<dynamic> pbSchemaDecoded = jsonDecode(pbSchemaFileContent);
+    final pbSchemaDecoded = jsonDecode(pbSchemaFileContent);
+    if (pbSchemaDecoded is! List<dynamic>) {
+      return;
+    }
 
     final collections = <Collection>[];
-    for (Map<String, dynamic> collectionJson in pbSchemaDecoded) {
+    for (final collectionJson in pbSchemaDecoded) {
+      if (collectionJson is! Map<String, dynamic>) {
+        continue;
+      }
+
       final collectionModel = Collection.fromJson(collectionJson);
       if (_generateSystemCollections || !collectionModel.system) {
         collections.add(collectionModel);
@@ -108,7 +113,7 @@ class Generator {
       geoPointClassGenerator('geo_point_class', _lineLength),
     );
 
-    for (var collection in collections) {
+    for (final collection in collections) {
       final fileName = '${ReCase(collection.name).snakeCase}_record';
 
       createFileAndWrite(