+++ lib/src/schema/collection/constructors/from_record_model.dart
@@ -19,7 +19,7 @@ code_builder.Constructor _fromRecordModelConstructor(String className) {
                 .declareFinal('extendedJsonMap')
                 .assign(code_builder.literalMap({
                   code_builder.literalSpread(): code_builder.refer('recordModel.data'),
-                  for (var recordFieldName in ['id', 'collectionId', 'collectionName'])
+                  for (final recordFieldName in ['id', 'collectionId', 'collectionName'])
                     code_builder.refer('${className}FieldsEnum.$recordFieldName.nameInSchema'):
                         code_builder.refer('recordModel.$recordFieldName'),
                 }))